rename tab components
