+++ packages/ui/src/lib/tabList/Introduction.mdx
@@ -1,0 +1,8 @@
+import { Meta } from '@storybook/blocks';
+
+<Meta title="Ui/TabList" />
+
+# TabList
+
+The `TabList` component provides a set of buttons for switching between tabs.
+It's styled in the way used by most CIU web maps/visualizations.
